Move `instamatic.typing` to `typing_` to avoid import conflicts
