+++ src/instamatic/microscope/base.py
@@ -4,7 +4,7 @@ from abc import ABC, abstractmethod
 from typing import Optional, Tuple
 
 from instamatic.microscope.utils import StagePositionTuple
-from instamatic.typing import float_deg, int_nm
+from instamatic.typing_ import float_deg, int_nm
 
 
 class MicroscopeBase(ABC):
