+++ src/instamatic/microscope/components/stage.py
@@ -8,7 +8,7 @@ import numpy as np
 
 from instamatic.microscope.base import MicroscopeBase
 from instamatic.microscope.utils import StagePositionTuple
-from instamatic.typing import float_deg, int_nm
+from instamatic.typing_ import float_deg, int_nm
 
 
 class Stage:
