+++ src/instamatic/microscope/interface/fei_microscope.py
@@ -12,7 +12,7 @@ from instamatic import config
 from instamatic.exceptions import FEIValueError, TEMCommunicationError
 from instamatic.microscope.base import MicroscopeBase
 from instamatic.microscope.utils import StagePositionTuple
-from instamatic.typing import float_deg, int_nm
+from instamatic.typing_ import float_deg, int_nm
 
 logger = logging.getLogger(__name__)
