+++ src/instamatic/microscope/interface/simu_microscope.py
@@ -8,7 +8,7 @@ from instamatic import config
 from instamatic.exceptions import TEMValueError
 from instamatic.microscope.base import MicroscopeBase
 from instamatic.microscope.utils import StagePositionTuple
-from instamatic.typing import float_deg, int_nm
+from instamatic.typing_ import float_deg, int_nm
 
 NTRLMAPPING = {
     'GUN1': 0,
